pyproject.toml
[project]
name = "zero-touch-site-assessor"
version = "0.1.0"
description = "Multimodal Zero-Touch Site Assessor for solar and heat pump installations"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Krish" }]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "google-genai>=1.0.0",
    "python-multipart>=0.0.9",
    "jsonschema>=4.23.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
strict = true
python_version = "3.12"

[tool.ruff]
target-version = "py312"
line-length = 100
